### 2018-03-06 #604 @whedon
[PRE REVIEW]: coalitions: Coalition probabilities in multi-party democracies
**Submitting author:** @adibender (<a href="http://orcid.org/0000-0001-5628-8611">Andreas Bender</a>)
**Repository:** <a href="https://github.com/adibender/coalitions" target ="_blank">https://github.com/adibender/coalitions</a>
**Version:** v0.6.0
