+++ R/scrappers.R
@@ -309,11 +309,13 @@ get_surveys_nds <- function() {
 #' @importFrom forcats fct_collapse
 #' @importFrom lubridate dmy
 #' @importFrom rlang UQS syms .data
+#' @importFrom stringr str_replace
 #' @export
 scrape_austria <- function(
   address = "https://neuwal.com/wahlumfragen/openwal/neuwal-openwal.json") {
 
-  aut_list <- fromJSON(getURL(address))
+  aut_list <- fromJSON(getURL(address) %>% 
+                           str_replace('\\"\\"(.*)\\"\\",', "\"'\\1'\",")) # fix for double double-quote bug
   out_df   <- as_tibble(aut_list) %>%
     rename(survey = "results") %>%
     select(one_of(c("institute", "date", "n",  "survey"))) %>%
