+++ man/get_surveys_nds.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/scrappers.R
+% Please edit documentation in R/scrapers.R
 \name{get_surveys_nds}
 \alias{get_surveys_nds}
 \title{Obtain (nested) Niedersachsen surveys object}
