+++ man/get_surveys_by.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/scrappers.R
+% Please edit documentation in R/scrapers.R
 \name{get_surveys_by}
 \alias{get_surveys_by}
 \title{Obtain (nested) Bavaria surveys object}
