+++ man/scrape_austria.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/scrappers.R
+% Please edit documentation in R/scrapers.R
 \name{scrape_austria}
 \alias{scrape_austria}
 \title{Import austrian survey results}
