+++ man/scrape.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/scrappers.R
+% Please edit documentation in R/scrapers.R
 \name{scrape_wahlrecht}
 \alias{scrape_wahlrecht}
 \alias{scrape_by}
