+++ man/extract_num.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/scrappers.R
+% Please edit documentation in R/scrapers.R
 \name{extract_num}
 \alias{extract_num}
 \title{Extract numericals from string or character}
