### 2018-03-09 #606 @leeper
@fsolt Thanks for your super fast review!! And props for submitting [a PR](https://github.com/adibender/coalitions/pull/79) to make fixes.
