### 2018-03-09 #606 @adibender
The whole review process was indeed super fast! Thank you all and especially @fsolt for that PR. Allready merged!
