### 2018-03-10 #606 @leeper
This looks good to me. @arfon over to you!
