### 2018-03-11 92d1462
Minor cosmetic updates before versioning for JOSS publication (after review)
