### 2018-03-11 cde4764
Update to V0.6.1
