+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: coalitions
 Type: Package
 Title: Coalition Probabilities in Multi-party Democracies
-Version: 0.6.0
+Version: 0.6.1
 Authors@R: c(
     person("Andreas", "Bender", , "andreas.bender@stat.uni-muenchen.de", role = c("aut", "cre"), comment=c(ORCID = "0000-0001-5628-8611")),
     person("Alexander", "Bauer", , "alexander.bauer@stat.uni-muenchen.de", role=c("aut"), comment=c(ORCID = "0000-0003-3495-5131"))
