### 2018-03-11 c9d776a
Cosmetic changes: Spell check update (found some more typos)
