+++ man/paste_coalitions.Rd
@@ -7,7 +7,7 @@
 paste_coalitions(coalitions, collapse = "_")
 }
 \arguments{
-\item{coalitions}{A list of coaltions for which coalition probabilities should
+\item{coalitions}{A list of coalitions for which coalition probabilities should
 be calculated. Each list entry must be a vector of party names. Those names
 need to correspond to the names in \code{majority_df}.}
 
