### 2018-03-11 #606 @adibender
@arfon  I archived on zenodo
- 10.5281/zenodo.1172594: This DOI represents all versions, and will always resolve to the latest one
- 10.5281/zenodo.1195667: This is the DOI of the current version

Not sure which one you usually use.
