+++ tests/testthat/test-scrapers.R
@@ -2,13 +2,13 @@ context("scrapers")
 
 test_that("State wide german scrapers work", {
 hp.vec <- c(
-	"allensbach"     = "http://www.wahlrecht.de/umfragen/allensbach.htm",
-	"emnid"          = "http://www.wahlrecht.de/umfragen/emnid.htm",
-	"forsa"          = "http://www.wahlrecht.de/umfragen/forsa.htm",
-	"forschGrWahlen" = "http://www.wahlrecht.de/umfragen/politbarometer.htm",
-	"gms"            = "http://www.wahlrecht.de/umfragen/gms.htm",
-	"infratest"      = "http://www.wahlrecht.de/umfragen/dimap.htm",
-	"insa"           = "http://www.wahlrecht.de/umfragen/insa.htm")
+	"allensbach"     = "https://www.wahlrecht.de/umfragen/allensbach.htm",
+	"emnid"          = "https://www.wahlrecht.de/umfragen/emnid.htm",
+	"forsa"          = "https://www.wahlrecht.de/umfragen/forsa.htm",
+	"forschGrWahlen" = "https://www.wahlrecht.de/umfragen/politbarometer.htm",
+	"gms"            = "https://www.wahlrecht.de/umfragen/gms.htm",
+	"infratest"      = "https://www.wahlrecht.de/umfragen/dimap.htm",
+	"insa"           = "https://www.wahlrecht.de/umfragen/insa.htm")
 
 expect_data_frame(head(scrape_wahlrecht(hp.vec[1])), ncols=11, nrows=6)
 expect_data_frame(head(scrape_wahlrecht(hp.vec[2])), ncols=11, nrows=6)
@@ -17,7 +17,7 @@ expect_data_frame(head(scrape_wahlrecht(hp.vec[4])), ncols=11, nrows=6)
 expect_data_frame(head(scrape_wahlrecht(hp.vec[5])), ncols=13, nrows=6)
 expect_data_frame(head(scrape_wahlrecht(hp.vec[6])), ncols=11, nrows=6)
 
-survey <- scrape_wahlrecht(address = "http://www.wahlrecht.de/umfragen/insa.htm" ) %>%
+survey <- scrape_wahlrecht(address = "https://www.wahlrecht.de/umfragen/insa.htm" ) %>%
 	filter(date==as.Date("2017-08-01"))
 expect_data_frame(survey, nrows=1, ncols=13)
 expect_equal(colnames(survey), c("date", "start", "end", "cdu", "spd",
@@ -25,7 +25,7 @@ expect_equal(colnames(survey), c("date", "start", "end", "cdu", "spd",
 expect_equal(survey$spd, 24.5)
 expect_equal(survey$respondents, 2046)
 
-survey2 <- scrape_wahlrecht(address = "http://www.wahlrecht.de/umfragen/allensbach.htm") %>%
+survey2 <- scrape_wahlrecht(address = "https://www.wahlrecht.de/umfragen/allensbach.htm") %>%
 	filter(date==as.Date("2017-07-18"))
 expect_data_frame(survey2, nrows=1, ncols=11)
 expect_equal(survey2$cdu, 39.5)
@@ -57,7 +57,7 @@ test_that("Federal german scrapers work", {
 
 context("Transform to long format")
 test_that("Collapse parties works correctly", {
-	surveys <- scrape_wahlrecht("http://www.wahlrecht.de/umfragen/gms.htm")[1,]
+	surveys <- scrape_wahlrecht("https://www.wahlrecht.de/umfragen/gms.htm")[1,]
 	surveys[, 5] <- NA
 	expect_data_frame(cp <- collapse_parties(surveys), nrows=1, ncols=5)
 })
