### 2018-03-11 #606 @arfon
@fsolt - many thanks for your review here and to @leeper for editing this one ✨
